pyproject.toml
[project]
name = "emotional-rag-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.1",
    "fastapi[standard]>=0.119.0",
    "google-genai>=1.45.0",
    "google-generativeai>=0.8.5",
    "sentence-transformers>=5.1.1",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "uvicorn>=0.37.0",
]
